pyproject.toml
[tool.poetry]
name = "file-to-markdown-conversion"
version = "0.1.0"
description = ""
authors = ["Ivan Korneichuk <korneichuk.ivan.ik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
tabulate = "^0.9.0"
markdownify = "^0.12.1"
pytesseract = "^0.3.10"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
